backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "sse-starlette>=3.1.2",
    "uvicorn>=0.40.0",
]
